#41: debug graph

+++ NERODevelopment/src/models/mock_model.h
@@ -73,7 +73,7 @@ public:
   std::optional<bool> getIsTalking() override;
   std::optional<int> getNumberOfCriticalFaults() override;
   std::optional<int> getNumberOfNonCriticalFaults() override;
-  std::optional<float> getTime() override;
+  std::optional<int> getTime() override;
 
   void sendMessage(const QString topic, const QString message) override;
